refactor: consolidate constants into 1 directory
