+++ src/components/BackButton/index.js
@@ -2,7 +2,7 @@
 import { jsx } from '@emotion/core';
 import { useHistory } from 'react-router-dom';
 
-import { Routes } from 'constants/Routes';
+import { Routes } from 'lib/constants/routes';
 
 import { IconButton } from 'components/Button';
